+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - MethodLogger moved to framework.db.local.logger package, now internal only.
 - TaskScheduler.Builder.TaskStep added.
 - State, overloaded builder() method removed.
+- ProxyBuilder.builder() renamed of().
 ### is.codion.common.db
 - Unit tests improved.
 - DatabaseConnection.getMethodLogger() now returns MethodLogger.noOpLogger() in case of no logger.